+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   fnToTable local function, converts array to html table for html info field
 -   GIT_CLONE_COMMAND and GIT_PULL_COMMAND env vars added
 -   fnLs dir,{recursive,regex}
+-   fnParseHtmlWithRegex url,regex,regexflags
 
 ### Fixed